Change temp data type from int16_t to uint16_t to eliminate overflow (#53)
* Change temp data type from int16_t to uint16_t to eliminate overflow

* Fix device drivers to use uint16_t instead of int data type.

Unsigned integers are used in the gc_board struct, however originally,
the functions passed signed intgers.

+++ src/gen_cranberry/gc_dev_hih6131.cpp
@@ -7,7 +7,7 @@ void gc_dev_hih6131_open(void){
     Wire.begin(9600);
 }
 
-int gc_dev_hih6131_temp_centik_read(void){
+uint16_t gc_dev_hih6131_temp_centik_read(void){
     int16_t value = 30000;
 
     #ifndef SEN_STUB
@@ -18,7 +18,7 @@ int gc_dev_hih6131_temp_centik_read(void){
     return value;
 }
 
-int gc_dev_hih6131_humidity_pct_read(void){
+uint16_t gc_dev_hih6131_humidity_pct_read(void){
     uint16_t value = 60;
 
     #ifndef SEN_STUB
